Slight changes to imports to allow alternative startup methods
For some reason, I had a little trouble with the original code organization
in my environment.  Python may possibly behave differently on Windows and
Macs with respect to how it searches the current directory.  The changes to
the import statements allow the following methods to work to start the
system:

1) run "run.py" directly from a shell

2) run "python3 -m guide"

These still allow arguments to work, so the --init argument should work
like before.